Use ftplugin instead of autocmds to set indent
The autocmds approach had the issue of overwriting the indentation
globally for all files. So if you opened a .nix file and then wanted to
edit a .c file, it would then still do 2 space indentation when it
should be 4. The autocmds approach was redundant in the first place as
the ftplugin directory already exists for this exact purpose and does it
without the aformentioned issue.